requirements.txt
streamlit==1.10.0
streamlit_option_menu==0.3.2
matplotlib==3.3.3
numpy==1.19.2
pandas==1.1.3
Pillow==9.4.0
Plost==0.2.5
plotly==4.14.1
scikit_learn==0.23.2
seaborn==0.10.0

